Exract and fix reveal tests
